+++ test/unit/fuzz-tools.test.js
@@ -408,6 +408,7 @@ import * as mrcPowerV428 from '../../lib/mrc-power-v428.js';
 import * as sarnatHieV429 from '../../lib/sarnat-hie-v429.js';
 import * as papileIvhV430 from '../../lib/papile-ivh-v430.js';
 import * as bellNecV431 from '../../lib/bell-nec-v431.js';
+import * as badenWalkerV432 from '../../lib/baden-walker-v432.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -761,6 +762,7 @@ const MODULES = {
   'sarnat-hie-v429.js': sarnatHieV429,
   'papile-ivh-v430.js': papileIvhV430,
   'bell-nec-v431.js': bellNecV431,
+  'baden-walker-v432.js': badenWalkerV432,
   'tb-testing.js': tbTesting,
 };